+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix live reload for Angular project with non default (app) project names
 - Fix running for iOS/Android to ensure stop option always available
+- Support for https with Nexus Browser
 
 ### Version 1.34.2