+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### Version 1.35
+
+- Fix live reload for Angular project with non default (app) project names
+- Fix running for iOS/Android to ensure stop option always available
+- Support for https with Nexus Browser
+
 ### Version 1.34.2
 
 - Fix for projects that appear to have folder based monorepos but are not